Added text input field to login form for username/email address
The original height used here was 20 which would not actually show the
text on an android device. It turns out that android devices need to
have a height twice that of what an iOS device needs so we used a
ternary statement to determine what height to use based on the device.