Modified the employee save to dispatch a success message to reducer
This makes sure that the reducer knows to reset the state of the
employee form so that it no longer shows the content for the employee
that was being edited prior to the firebase save succeeding. If this was
not done, the next time the employee create form was opened, the last
employee that was opened using the edit form would show.